Add files via upload
The Stochastic Player sends midiout data for a selection of channels. It is recommended that you connect this up with a DAW to receive the midi data and select software instruments for each of the 4 channels. See playlist of video tutorial on youtube:  https://www.youtube.com/watch?v=m7iNPb5XnHM&list=PL2hVNB6jE4P20CKyCo-BiB_NOERIlWy3c